requirements.txt
Flask==3.0.0
flask-cors==4.0.0
neo4j==5.15.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Jinja2==3.1.2
Werkzeug==3.0.1
pydantic==2.5.0
